Add PsyZ submodule on CI (#199)
This allows stubbing the PSX SDK includes with PsyZ ones when verifying
the source code against a modern compiler. Required as we started adding
GTE macros in the psxsdk headers.

+++ src/world/world.c
@@ -3523,7 +3523,7 @@ static void* WmGetSkyboxOverlayCurrTextureSettingBuffer(void) { return (WmGetCur
 
 static s32 func_800AE628(void) { return D_8010B080; }
 
-void WmUpdateSkyboxOverlayVertexes(s32 angle) {
+void WmUpdateSkyboxOverlayVertexes(s16 angle) {
     SVECTOR v;
     SVECTOR rot;
     VECTOR out;
